requirements.txt
Flask==2.3.1
flake8==6.0.0
pytest==7.3.1
pytest-cov==4.0.0
isort==5.12.0
safety==2.3.5
Flask-Login==0.6.2
email-validator==2.0.0.post2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
bandit==1.7.5
pydantic==1.10.7
